Rename bakeCovidExplorer.ts → bakeCovidExplorerRedirects.ts
