+++ site/server/bakeCovidExplorerRedirects.ts
